NPNTool.NPNConstr

Datatypes

data NPNConstr l v c

Datatype for dynamically constructing Nested Petri nets

Constructors

NPNConstr

Fields

p :: Map PTPlace [Either c ElemNetId]

key :: Int

keyT :: Int

keyEN :: Int

tin :: Map Trans (SArc v c PTPlace)

tout :: Map Trans (SArc v c PTPlace)

nets :: IntMap (ElemNet l)

tlab :: Map Trans l

type NPNConstrM l v = State (NPNConstr l v Int)

Operations

new :: NPNConstr l v c

New (empty) NPNConstr

Monadic interface

run :: (Ord v, Ord l) => NPNConstrM l v a -> NPNConstr l v Int -> (a, NPNet l v Int)

Runs a NPNConstrM monad and returns a NP-net together with a result

addElemNet :: ElemNet l -> NPNConstrM l v ElemNetId

Adds an existing element net to the system

mkPlace :: Ord v => NPNConstrM l v PTPlace

Creates a new place not yet present in the net

mkTrans :: Ord v => NPNConstrM l v Trans

Creates a new tranisition not yet present in the net

label :: Ord v => Trans -> l -> NPNConstrM l v ()

Specifies a label for some transition

insPlace :: Ord v => PTPlace -> NPNConstrM l v ()

Inserts a place into the system

inT :: Ord v => PTPlace -> Expr v Int -> Trans -> NPNConstrM l v ()

outT :: Ord v => Trans -> Expr v Int -> PTPlace -> NPNConstrM l v ()

mark :: PTPlace -> Either Int ElemNetId -> NPNConstrM l v ()

Marks a place with a token or an element net

marks :: PTPlace -> [Either Int ElemNetId] -> NPNConstrM l v ()

Marks a place with several tokens or element nets

PTConstrM and PTNet interface

liftPTC :: Ord v => PTConstrM l a -> NPNConstrM l v a

Lift a PTConstrM constructions to NPNConstrM monad

liftElemNet :: PTConstrM l a -> NPNConstrM l v ElemNetId

Lifts an elementary net, described by the PTConstrM into the NPNConstrM monad and adds it to the system.

insElemNet :: Int -> PTConstrM l a -> NPNConstrM l v a

Useful for working with a priori correct data, generally use addElemNet instead

Generalized arcs (with expressions)

class Arc a => ArcExpr a v where

Methods

arcExpr :: a -> Expr v Int -> Co a -> NPNConstrM l v ()

arc :: a -> Co a -> NPNConstrM l v ()