NPNTool.PTConstr

Datatypes

data PTConstr l

Datatype for dynamically constructing Petri Nets

Constructors

PTConstr

Fields

p :: Map PTPlace Int

key :: Int

keyT :: Int

tin :: Map Trans (MultiSet PTPlace)

tout :: Map Trans (MultiSet PTPlace)

tlab :: Map Trans l

type PTConstrM l = State (PTConstr l)

Operations

new :: forall l. PTConstr l

New (empty) PTConstr

Monadic interface

run :: PTConstrM l a -> PTConstr l -> (a, PTNet)

Runs a PTConstrM monad and returns a Petri Net together with a result

runL :: PTConstrM l a -> PTConstr l -> (a, PTNet, Labelling l)

Runs a PTConstrM monad and returns a Petri Net together with its labelling and a result

runConstr :: PTConstrM l a -> PTConstr l -> PTConstr l

Runs a PTConstrM monad and returns a PTConstr

mkPlace :: PTConstrM l PTPlace

Creates a new place not yet present in the net

insPlace :: PTPlace -> PTConstrM l ()

Inserts an existing place in the net Does nothing if the place is already present

mkTrans :: PTConstrM l Trans

Creates a new transition not yet present in the net

label :: Trans -> l -> PTConstrM l ()

Specifies a label for some transition

mark :: PTPlace -> PTConstrM l ()

Marks a place with a token

marking :: [PTPlace] -> PTConstrM l ()

Marks a set of places

inT :: PTPlace -> Trans -> PTConstrM l ()

outT :: Trans -> PTPlace -> PTConstrM l ()

inTn :: PTPlace -> Trans -> Int -> PTConstrM l ()

outTn :: Trans -> PTPlace -> Int -> PTConstrM l ()

Generalized arcs

class Arc k where

Generalized arc class

Associated Types

type Co k :: *

Methods

arc :: k -> Co k -> PTConstrM l ()

Specifies an arc from some object to its co-object

arcn :: Arc k => k -> Co k -> Int -> PTConstrM l ()

Specifies n arcs

conn :: Show l => PTPlace -> PTPlace -> l -> PTConstrM l ()

Directly connects two places via an intermediate transition Might be useful for specifying workflow nets