NPNTool.PetriNet

Datatypes

data Net p t n m

A generalized net datatype

Constructors

Net

Fields

places :: Set p

A set of places

trans :: Set t

A set of transitions

pre :: t -> n p

The pre function

post :: t -> n p

The post function

initial :: m p

Initial marking

newtype Trans

Constructors

Trans

Fields

name :: String

type SS = Gr PTMark PTTrans

class DynNet net p t m | net -> p, net -> t, net -> m where

A general dynamic net interface

Methods

enabledS :: net -> m p -> Set t -> Bool

Whether a set of transitions is enabled. Note that this is different then checking whether each transition in the set is enabled

enabled :: net -> m p -> t -> Bool

Whether some transition is enabled

fire :: net -> m p -> t -> m p

The marking after some transitions is fired

type PTNet = Net PTPlace Trans MultiSet MultiSet

type PTMark = MultiSet PTPlace

type PTTrans = Trans

type PTPlace = Int

General and abstract functions, opertations

data HLArc a p

type LLArc = HLArc Identity

annotate :: (Functor a, Functor n) => Net p t n m -> (p -> t -> a p) -> (t -> p -> a p) -> Net p t (Compose n a) m

Annotate net using additional functors of place/transition relationship

Additional functions

fireSequence_ :: DynNet net p t m => net -> m p -> [t] -> m p

fireSequence :: DynNet net p t m => net -> m p -> [t] -> Maybe (m p)

reachabilityGraph :: (DynNet (Net p t n m) p t m, Ord p, Ord (m p)) => Net p t n m -> Gr (m p) t

The reachability graph of a Petri Net

reachabilityGraph' :: (DynNet (Net p t n m) p t m, Ord p, Ord (m p)) => Net p t n m -> (NodeMap (m p), Gr (m p) t)

The reachability graph of a Petri Net together with a NodeMap

postP :: (Eq p, Foldable n) => p -> Net p t n m -> [t]

Post-transitions of a place

preP :: (Eq p, Foldable n) => p -> Net p t n m -> [t]

Pre-transitions of a place