NPNTool.Unfoldings

Description

Petri net unfoldings

Occurence nets and branching processes

type Event = PTTrans

type Condition = PTPlace

type OccurNet = Net Condition Event Set Set

Occurrence net

type Hom = (Condition -> PTPlace, Event -> PTTrans)

Homomorphism from P/T-net to occurence net

type BProc = (OccurNet, Hom)

Branching process

toOccurNet :: PTConstr l -> OccurNet

Configurations and cuts

type Conf = Set Event

A configuration of the branching process is a set of causally closed and conflict-free events

type Cut = Set Condition

cut :: BProc -> Conf -> Cut

cutMark :: BProc -> Cut -> MultiSet PTPlace

localConf :: BProc -> Event -> Conf

Orders and relations on nodes of branching processes

class Node n m

Associated Types

type CoNode n

predPred :: (Ord n, Node n OccurNet, Node (CoNode n) OccurNet, n ~ CoNode (CoNode n)) => BProc -> n -> Set n

Predecessor nodes of the same type

predR :: (Ord n, Node n OccurNet, Node (CoNode n) OccurNet, n ~ CoNode (CoNode n)) => BProc -> n -> Set n

Reflexive predecessor relationship

preds :: (Ord n, Node n OccurNet, Node (CoNode n) OccurNet, n ~ CoNode (CoNode n)) => BProc -> Set n -> Set n

Transitive closure of the predecessor relationship

predTrans :: BProc -> Condition -> Set Event

conflict :: BProc -> Condition -> Condition -> Bool

Whether two places are in conflict

causalPred :: BProc -> Condition -> Condition -> Bool

Whether one place is a causal predecessor of another

concurrent :: BProc -> Condition -> Condition -> Bool

Whether two places are concurrent

Unfoldings and prefixes

type EventQueue = [(PTTrans, Set Condition)]

posTrans :: BProc -> PTNet -> PTTrans -> Maybe (Set Condition)

Whether it is possbile to add a transition to a branching process. Return Nothing if it's not, returns Just s otherwise, where s is a pre-set of the transition which should be added.

posExt :: BProc -> PTNet -> Set Event -> EventQueue

unfolding :: Int -> PTNet -> BPConstrM ()

Run the unfolding algorithm for k steps or until it stops

Auxiliary functions

pairs :: [a] -> [a] -> [(a, a)]

fixedPoint :: Eq a => (a -> a) -> a -> a

General fixed point funciton since the fix from Data.Function only compute Kleene fixed-points

choose :: Int -> [a] -> [[a]]

All the possible ways to choose k elements from a list

type BPConstrM = StateT HomS (PTConstrM PTTrans)

runBPConstrM :: BPConstrM a -> (a, BProc)